+++ lib/msg91.dart
@@ -1,4 +1,7 @@
 /// Msg91 Official API Wrapper (Unofficial Dart Package)
 library msg91;
 
+export 'src/exceptions.dart';
 export 'src/msg91_base.dart';
+export 'src/msg91_otp.dart';
+export 'src/msg91_sms.dart';