+++ lib/src/exceptions.dart
@@ -1,0 +1,5 @@
+class Msg91Exception implements Exception {
+  String message;
+
+  Msg91Exception(this.message);
+}